restructured the folders to add pages for witness statement and other people's witness statement
