+++ e2e/fixtures/citizen/caseView/uploadDocuments/witnessStatementsAndEvidence/courtPermissionContent.ts
@@ -1,0 +1,7 @@
+export enum CourtPermissionContent {
+  GovukCaptionL = "Witness statements and evidence",
+  GovukHeadingXLPositionStatement = "Your position statement",
+  GovukHeadingXLWitnessStatement = "Your witness statement",
+  GovukHeadingXLOtherPeopleWitnessStatement = "Other people's witness statements",
+  GovukFieldsetLegend = "Has the court asked for this document?",
+}
