+++ e2e/fixtures/citizen/caseView/uploadDocuments/witnessStatementsAndEvidence/documentSharingDetailsContent.ts
