--- e2e/fixtures/citizen/caseView/uploadDocuments/yourPositionStatement/courtPermissionContent.ts
@@ -1,5 +1,0 @@
-export enum CourtPermissionContent {
-  GovukCaptionL = "Witness statements and evidence",
-  GovukHeadingXL = "Your position statement",
-  GovukFieldsetLegend = "Has the court asked for this document?",
-}
