+++ e2e/pages/citizen/caseView/uploadDocuments/witnessStatementsAndEvidence/documentSharingDetailsPage.ts
@@ -1,0 +1,61 @@
+import { Page } from "@playwright/test";
+import { Selectors } from "../../../../../common/selectors.ts";
+import { CommonStaticText } from "../../../../../common/commonStaticText.ts";
+import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper.ts";
+import { DocumentSharingDetailsContent } from "../../../../../fixtures/citizen/caseView/uploadDocuments/witnessStatementsAndEvidence/documentSharingDetailsContent.ts";
+import { Helpers } from "../../../../../common/helpers.ts";
+
+export class DocumentSharingDetailsPage {
+  public static async documentSharingDetailsPage(
+    page: Page,
+    accessibilityTest: boolean,
+  ): Promise<void> {
+    // this part of the case is complete when the case is created through courtnav so only need to check the page
+    await this.checkPageLoads(page, accessibilityTest);
+    await this.continue(page);
+  }
+
+  private static async checkPageLoads(
+    page: Page,
+    accessibilityTest: boolean,
+  ): Promise<void> {
+    await page
+      .locator(Selectors.GovukHeadingL, {
+        hasText: DocumentSharingDetailsContent.GovukHeadingL,
+      })
+      .waitFor();
+    await Promise.all([
+      Helpers.checkVisibleAndPresent(
+        page,
+        `${Selectors.GovukHeadingM}:text-is("${DocumentSharingDetailsContent.GovukHeadingM}")`,
+        1,
+      ),
+      Helpers.checkVisibleAndPresent(
+        page,
+        `${Selectors.GovukBody}:text-is("${DocumentSharingDetailsContent.GovukBody1}")`,
+        1,
+      ),
+      Helpers.checkVisibleAndPresent(
+        page,
+        `${Selectors.GovukLink}:text-is("${DocumentSharingDetailsContent.GovukLink}")`,
+        1,
+      ),
+      Helpers.checkVisibleAndPresent(
+        page,
+        `${Selectors.GovukBody}:text-is("${DocumentSharingDetailsContent.GovukBody2}")`,
+        1,
+      ),
+    ]);
+    if (accessibilityTest) {
+      await AccessibilityTestHelper.run(page);
+    }
+  }
+
+  private static async continue(
+    page: Page,
+  ): Promise<void> {
+    await page.click(
+      `${Selectors.GovukButton}:text-is("${CommonStaticText.continue}")`,
+    );
+  }
+}
