+++ e2e/pages/citizen/caseView/uploadDocuments/witnessStatementsAndEvidence/positionStatementPage.ts
@@ -2,17 +2,24 @@ import { Page } from "@playwright/test";
 import { Selectors } from "../../../../../common/selectors.ts";
 import { CommonStaticText } from "../../../../../common/commonStaticText.ts";
 import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper.ts";
-import { SubmitExtraEvidenceContent } from "../../../../../fixtures/citizen/caseView/uploadDocuments/yourPositionStatement/submitExtraEvidenceContent.ts";
+import { CourtPermissionContent } from "../../../../../fixtures/citizen/caseView/uploadDocuments/witnessStatementsAndEvidence/courtPermissionContent.ts";
 import { Helpers } from "../../../../../common/helpers.ts";
+import { yesNoNA } from "../../../../../common/types.ts";
 
-export class SubmitExtraEvidencePage {
-  public static async submitExtraEvidencePage(
+enum UniqueSelectors {
+  courtHasAskedForDoc = "#hasCourtAskedForThisDoc",
+  courtHasNotAskedForDoc = "#hasCourtAskedForThisDoc-2",
+}
+
+export class PositionStatementPage {
+  public static async courtPermissionPage(
     page: Page,
     accessibilityTest: boolean,
+    yesNoNA: yesNoNA,
   ): Promise<void> {
     // this part of the case is complete when the case is created through courtnav so only need to check the page
     await this.checkPageLoads(page, accessibilityTest);
-    await this.continue(page);
+    await this.fillInFields(page, yesNoNA);
   }
 
   private static async checkPageLoads(
@@ -20,24 +27,19 @@ export class SubmitExtraEvidencePage {
     accessibilityTest: boolean,
   ): Promise<void> {
     await page
-      .locator(Selectors.GovukHeadingL, {
-        hasText: SubmitExtraEvidenceContent.GovukHeadingL,
+      .locator(Selectors.GovukHeadingXL, {
+        hasText: CourtPermissionContent.GovukHeadingXLPositionStatement,
       })
       .waitFor();
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukCaptionL}:text-is("${SubmitExtraEvidenceContent.GovukCaptionL}")`,
+        `${Selectors.GovukCaptionL}:text-is("${CourtPermissionContent.GovukCaptionL}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukBody}:text-is("${SubmitExtraEvidenceContent.GovukBody1}")`,
-        1,
-      ),
-      Helpers.checkVisibleAndPresent(
-        page,
-        `${Selectors.GovukBody}:text-is("${SubmitExtraEvidenceContent.GovukBody2}")`,
+        `${Selectors.GovukFieldsetLegend}:text-is("${CourtPermissionContent.GovukFieldsetLegend}")`,
         1,
       ),
     ]);
@@ -46,9 +48,17 @@ export class SubmitExtraEvidencePage {
     }
   }
 
-  private static async continue(
+  private static async fillInFields(
     page: Page,
+    yesNoNA: yesNoNA,
   ): Promise<void> {
+    if (yesNoNA === "Yes") {
+      await page.click(UniqueSelectors.courtHasAskedForDoc);
+    } else if (yesNoNA === "No") {
+      await page.click(UniqueSelectors.courtHasNotAskedForDoc);
+    } else {
+      throw new Error("Invalid value for yesNoNA");
+    }
     await page.click(
       `${Selectors.GovukButton}:text-is("${CommonStaticText.continue}")`,
     );
